refactor: improve API compatibility and file structure organization
- Fix HCOMM_CHANNEL_MAGIC_WORD value (0x010f0f0f -> 0x0f0f0f0f)
- Refactor type definitions in hcomm_res_defs.h:
  - Convert COMM_ADDR_EID_LEN to enum for better type safety
  - Change HcommResult from enum to int32_t typedef
  - Add HcclMemType enum for memory type compatibility
- Move hcomm_primitives_inner.h to proper location under data_api directory
- Update CMakeLists.txt to handle header installation with symbolic links
- Remove obsolete scripts (cleancode_check.sh, install_gemini_cli.sh, sync-ai-system-skills-bridge.js)
- Update CommLib.xml module configuration
- Fix header includes and references across framework modules
- Update unit tests to work with refactored type definitions

+++ test/ut/framework/next/comms/comm_engine_res/threads/test_thread_api.cc
@@ -204,7 +204,8 @@ TEST_F(TestHcclThread, Ut_TestHcommThreadAlloc_When_Allocate_MAXThreadNum_expect
     .with(outBound(DevType::DEV_TYPE_950))
     .will(returnValue(HCCL_SUCCESS)); 
     ThreadHandle thread;
-    HcommResult ret =  HcommThreadAlloc(COMM_ENGINE_CPU_TS, hccl::HCOMM_THREADNUM_MAX_NUM + 1, 0, &thread);
+    HcommResult ret = HcommThreadAlloc(
+        COMM_ENGINE_CPU_TS, hccl::HCOMM_THREADNUM_MAX_NUM + 1, static_cast<uint32_t>(0), &thread);
     EXPECT_EQ(ret, HCCL_E_PARA);
 }
